# API/src/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Fitness API",
    "description": "A comprehensive fitness tracking and coaching API that allows coaches to manage trainees, create workout and nutrition plans, and track progress.",
    "contact": {
      "name": "API Support",
      "email": "support@fitness-api.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for authentication"
      }
    },
    "schemas": {
...
      "SuccessMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Success message",
            "example": "Operation completed successfully"
          }
        },
        "required": [
          "message"
        ]
      },
      "ConnectRequest": {
        "type": "object",
        "properties": {
          "coachId": {
            "type": "string",
            "minLength": 1,
            "description": "ID of the coach to connect with",
            "example": "coach123"
          },
          "notes": {
            "type": "string",
            "description": "Optional notes for connection request",
            "example": "Looking for strength training guidance"
          }
        },
        "required": [
          "coachId"
        ]
      },
...
      "TraineeId": {
        "type": "object",
        "properties": {
          "traineeId": {
            "type": "string",
            "minLength": 1,
            "description": "ID of the trainee",
            "example": "trainee456"
          }
        },
        "required": [
          "traineeId"
        ]
      },
...
      "Connection": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Connection ID",
            "example": 1
          },
          "coachId": {
            "type": "string",
            "description": "Coach user ID"
          },
          "traineeId": {
            "type": "string",
            "description": "Trainee user ID"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "accepted",
              "rejected",
              "ended"
            ],
            "description": "Connection status"
          },
          "notes": {
            "type": "string",
            "description": "Connection notes"
          },
          "requestedAt": {
            "type": "string",
            "description": "Request date"
          },
          "respondedAt": {
            "type": "string",
            "description": "Response date"
          }
        },
        "required": [
          "id",
          "coachId",
          "traineeId",
          "status",
          "requestedAt"
        ]
      },
      "ExerciseType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Exercise type ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Exercise name",
            "example": "Push-ups"
          },
          "devicePosition": {
            "type": "string",
            "enum": [
              "thigh",
              "arm",
              "none"
            ],
            "description": "Device position for tracking",
            "example": "thigh"
          },
          "logType": {
            "type": "string",
            "enum": [
              "reps",
              "duration"
            ],
            "description": "Type of logging for this exercise",
            "example": "reps"
          }
        },
        "required": [
          "id",
          "name",
          "devicePosition",
          "logType"
        ]
      },
      "RecordExerciseResult": {
        "type": "object",
        "properties": {
          "workoutPlanDayExerciseId": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0
              },
              {
                "type": "string"
              }
            ],
            "description": "ID of the workout plan day exercise",
            "example": 789
          },
          "reps": {
            "type": "number",
            "exclusiveMinimum": 0,
            "description": "Number of repetitions completed",
            "example": 15
          },
          "duration": {
            "type": "number",
            "exclusiveMinimum": 0,
            "description": "Duration of exercise in seconds",
            "example": 180
          },
          "calories": {
            "type": "number",
            "exclusiveMinimum": 0,
            "description": "Calories burned during exercise",
            "example": 45
          }
        },
        "required": [
          "workoutPlanDayExerciseId"
        ]
      },
...
      "ExerciseResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Result ID"
          },
          "workoutPlanDayExerciseId": {
            "type": "number"
          },
          "userId": {
            "type": "string"
          },
          "reps": {
            "type": "number"
          },
          "duration": {
            "type": "number"
          },
          "calories": {
            "type": "number"
          },
          "recordedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "workoutPlanDayExerciseId",
          "userId",
          "recordedAt"
        ]
      },
      "ExerciseResultWithDetails": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Result ID"
          },
          "workoutPlanDayExerciseId": {
            "type": "number"
          },
          "userId": {
            "type": "string"
          },
          "reps": {
            "type": "number"
          },
          "duration": {
            "type": "number"
          },
          "calories": {
            "type": "number"
          },
          "completedAt": {
            "type": "string"
          },
          "workoutPlanExercise": {
            "type": "object",
            "properties": {
              "exerciseType": {
                "$ref": "#/components/schemas/ExerciseType"
              },
              "workoutPlanDay": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number"
                  },
                  "day": {
                    "type": "number"
                  },
                  "isRestDay": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "id",
                  "day",
                  "isRestDay"
                ]
              }
            },
            "required": [
              "exerciseType",
              "workoutPlanDay"
            ]
          }
        },
        "required": [
          "id",
          "workoutPlanDayExerciseId",
          "userId",
          "completedAt",
          "workoutPlanExercise"
        ]
      },
      "ExerciseResultForPlanExercise": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Result ID"
          },
          "workoutPlanDayExerciseId": {
            "type": "number"
          },
          "userId": {
            "type": "string"
          },
          "reps": {
            "type": "number"
          },
          "duration": {
            "type": "number"
          },
          "calories": {
            "type": "number"
          },
          "completedAt": {
            "type": "string"
          },
          "workoutPlanExercise": {
            "type": "object",
            "properties": {
              "exerciseType": {
                "$ref": "#/components/schemas/ExerciseType"
              }
            },
            "required": [
              "exerciseType"
            ]
          }
        },
        "required": [
          "id",
          "workoutPlanDayExerciseId",
          "userId",
          "completedAt",
          "workoutPlanExercise"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/exercises": {
      "get": {
        "tags": [
          "Exercises"
        ],
        "summary": "Get all exercise types",
        "description": "Retrieve a list of all available exercise types",
        "responses": {
          "200": {
            "description": "List of exercise types",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExerciseType"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/exercises/{id}": {
      "get": {
        "tags": [
          "Exercises"
        ],
        "summary": "Get exercise type by ID",
        "description": "Retrieve a specific exercise type by its ID",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d+$",
              "description": "Unique identifier",
              "example": "123"
            },
            "required": true,
            "description": "Unique identifier",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Exercise type details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseType"
                }
              }
            }
          },
          "404": {
            "description": "Exercise type not found"
          }
        }
      }
    },
    "/api/exercises/name/{name}": {
      "get": {
        "tags": [
          "Exercises"
        ],
        "summary": "Get exercise type by name",
        "description": "Retrieve a specific exercise type by its name",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Name of the exercise",
              "example": "Push-ups"
            },
            "required": true,
            "description": "Name of the exercise",
            "name": "name",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Exercise type details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseType"
                }
              }
            }
          },
          "404": {
            "description": "Exercise type not found"
          }
        }
      }
    },
    "/api/exercises/exercise-results": {
      "post": {
        "tags": [
          "Exercises"
        ],
        "summary": "Record exercise result",
        "description": "Record the result of an exercise performance",
        "requestBody": {
          "description": "Exercise result data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordExerciseResult"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Exercise result recorded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          }
        }
      }
    },
    "/api/exercises/results": {
      "get": {
        "tags": [
          "Exercises"
        ],
        "summary": "Get user exercise results",
        "description": "Retrieve all exercise results for the authenticated user",
        "responses": {
          "200": {
            "description": "List of user exercise results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExerciseResultWithDetails"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/exercises/workout-plan-day-exercise/{id}": {
      "get": {
        "tags": [
          "Exercises"
        ],
        "summary": "Get exercise results for workout plan day exercise",
        "description": "Retrieve exercise results for a specific workout plan day exercise",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d+$",
              "description": "Unique identifier",
              "example": "123"
            },
            "required": true,
            "description": "Unique identifier",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Exercise results for the specified workout plan day exercise",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExerciseResultForPlanExercise"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/connections/connect": {
      "post": {
        "tags": [
          "Connections"
        ],
        "summary": "Send connection request",
        "description": "Send a connection request from trainee to coach",
        "requestBody": {
          "description": "Connection request data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Connection request sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Connection"
                }
              }
            }
          }
        }
      }
    },
    "/api/connections/requests/{type}": {
      "get": {
        "tags": [
          "Connections"
        ],
        "summary": "Get connection requests",
        "description": "Get connection requests (sent or received)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "sent",
                "received"
              ],
              "description": "Type of connection requests to retrieve",
              "example": "sent"
            },
            "required": true,
            "description": "Type of connection requests to retrieve",
            "name": "type",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "List of connection requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Connection"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/api/connections/accept": {
      "post": {
        "tags": [
          "Connections"
        ],
        "summary": "Accept connection request",
        "description": "Accept a connection request (coach only)",
        "requestBody": {
          "description": "Trainee ID to accept",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TraineeId"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Connection request accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Connection"
                }
              }
            }
          }
        }
      }
    },
    "/api/connections/reject": {
      "post": {
        "tags": [
          "Connections"
        ],
        "summary": "Reject connection request",
        "description": "Reject a connection request (coach only)",
        "requestBody": {
          "description": "Trainee ID to reject",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TraineeId"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Connection request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessage"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}